backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
supabase==2.13.0
anthropic==0.52.0
sentence-transformers==4.1.0
Pillow==11.2.1
torch==2.6.0
numpy==2.2.4
httpx==0.28.1
